Display images if the question text is an image.
Currently, we detect if the question text is an image by checking the
extension.  We should probably instead check if it display in an <img>
tag, assuming that there are no security issues with this.  We should
possibly make this a configurable option.

I don't like permitting all network resources, especially because I
might forget to add something to the cache manifest.